Fix handling of dialogs on click
Example:
There is a button, which raises confirmation window on click.

Handling in page object is like:
```
$button->press();
$this->getDriver()->acceptAlert();
```

This is related to #18 - I found that initial solution is not comprehensive as it seems to be sending "wait" message to browser at least once even while it is locked with dialog.

The breaking flow:
```
-> Input.dispatchMouseEvent
<- Page.javascriptDialogOpening
-> Runtime.evaluate (wait)
<- dies and never sends handling command
```

Fix alters flow like this:
1. add back the condition not to wait if dialog is open
2. add the wait in acceptAlert/denyAlert after handling dialog

Then the flow is:
```
-> Input.dispatchMouseEvent
<- Page.javascriptDialogOpening
-> Page.handleJavaScriptDialog
<- Page.javascriptDialogClosed
-> Runtime.evaluate (wait)
```